fdss

+++ app/Controllers/Pdt0621.php
@@ -127,10 +127,12 @@ class Pdt0621 extends BaseController
                     return $this->response->setJSON(['status' => 'error', 'message' => "El archivo de renta no es correcto, RUC no coincide.", 'datos' => $datos_pdt_file]);
                 }
 
-                if ($datos_pdt_file['periodo'] != $ani . $data_periodo['mes_fecha']) {
-                    unlink(FCPATH . '/archivos/pdt/' . $archivo_pdt);
-                    unlink(FCPATH . '/archivos/pdt/' . $archivo_constancia);
-                    return $this->response->setJSON(['status' => 'error', 'message' => "El archivo de renta no es correcto, periodo no coincide.", 'datos' => $datos_pdt_file]);
+                if ($datos_pdt_file['periodo'] != null) {
+                    if ($datos_pdt_file['periodo'] != $ani . $data_periodo['mes_fecha']) {
+                        unlink(FCPATH . '/archivos/pdt/' . $archivo_pdt);
+                        unlink(FCPATH . '/archivos/pdt/' . $archivo_constancia);
+                        return $this->response->setJSON(['status' => 'error', 'message' => "El archivo de renta no es correcto, periodo no coincide.", 'datos' => $datos_pdt_file]);
+                    }
                 }
             }
